+++ R/comparedf.control.R
@@ -1,7 +1,7 @@
 
-#' Control settings for `compare` function
+#' Control settings for \code{comparedf} function
 #'
-#' Control tolerance definitions for the \code{\link{compare.data.frame}} function.
+#' Control tolerance definitions for the \code{\link{comparedf}} function.
 #'
 #' @param tol.logical,tol.num,tol.char,tol.factor,tol.date,tol.other A function or one of the shortcut character strings,
 #'   denoting the tolerance function to use for a given data type. See "details", below.
@@ -9,10 +9,12 @@
 #' @param int.as.num Logical; should integers be coerced to numeric before comparison? Default FALSE.
 #' @param factor.as.char Logical; should factors be coerced to character before comparison? Default FALSE.
 #' @param tol.date.val Numeric; maximum value of differences allowed in dates (fed to the function given in \code{tol.date}).
-#' @param tol.vars Either \code{"none"} (the default), denoting that variable names are to be matched as-is, or a
+#' @param tol.vars Either \code{"none"} (the default), denoting that variable names are to be matched as-is,
+#'   a named vector manually specifying variable names to compare (where the names correspond to columns of
+#'   \code{x} and the values correspond to columns of \code{y}), or a
 #'   character vector denoting equivalence classes for characters in the variable names. See "details", below.
 #' @param ... Other arguments (not in use at this time).
-#' @return A list containing the necessary parameters for the \code{\link{compare.data.frame}} function.
+#' @return A list containing the necessary parameters for the \code{\link{comparedf}} function.
 #' @details
 #' The following character strings are accepted:
 #' \itemize{
@@ -30,7 +32,8 @@
 #'   \item{\code{tol.other = "none"}: expect objects of other classes to be exactly identical.}
 #' }
 #'
-#' \code{tol.vars}: If not set to \code{"none"} (the default), the \code{tol.vars} argument is a character vector denoting equivalence classes
+#' \code{tol.vars}: If not set to \code{"none"} (the default) or a named vector,
+#'   the \code{tol.vars} argument is a character vector denoting equivalence classes
 #'   for the characters in the variable names. A single character in this vector means to replace that character
 #'   with \code{""}. All other strings in this vector are split by character and replaced by the first character in the string.
 #'
@@ -39,16 +42,16 @@
 #'
 #' The special character string \code{"case"} in this vector is the same as specifying \code{paste0(letters, LETTERS)}.
 #' @examples
-#' cntl <- comparison.control(
+#' cntl <- comparedf.control(
 #'   tol.num = "pct",     # calculate percent differences
 #'   tol.vars = c("case", # ignore case
 #'                "._",   # set all underscores to dots.
 #'                "e")    # remove all letter e's
 #' )
-#' @seealso \code{\link{compare.data.frame}}, \code{\link{comparison.tolerances}}
+#' @seealso \code{\link{comparedf}}, \code{\link{comparedf.tolerances}}, \code{\link{summary.comparedf}}
 #' @author Ethan Heinzen
 #' @export
-comparison.control <- function(
+comparedf.control <- function(
   tol.logical = "none",
   tol.num = c("absolute", "percent", "pct"),
   tol.num.val = sqrt(.Machine$double.eps),
@@ -83,11 +86,14 @@ comparison.control <- function(
   if(!is.function(tol.other)) tol.other <- match.fun(paste0("tol.other.", match.arg(tol.other, several.ok = FALSE)))
 
   #### Variable names ####
-  if(!is.character(tol.vars)){stop("'tol.vars' must be a character string or vector.")}
-  if("none" %in% tol.vars || length(tol.vars) == 0) tol.vars <- "none"
-  if("case" %in% tol.vars) tol.vars <- c(paste0(letters, LETTERS), tol.vars[tol.vars != "case"])
+  if(!is.character(tol.vars)) stop("'tol.vars' must be a character string or vector.")
+  if(is.null(names(tol.vars)))
+  {
+    if("none" %in% tol.vars || length(tol.vars) == 0) tol.vars <- "none"
+    if("case" %in% tol.vars) tol.vars <- c(paste0(letters, LETTERS), tol.vars[tol.vars != "case"])
+  }
 
-  return(list(tol.logical = tol.logical, tol.num = tol.num, tol.num.val = tol.num.val, int.as.num = int.as.num, tol.char = tol.char,
-              tol.factor = tol.factor, factor.as.char = factor.as.char, tol.date = tol.date, tol.date.val = tol.date.val,
-              tol.other = tol.other, tol.vars = tol.vars))
+  list(tol.logical = tol.logical, tol.num = tol.num, tol.num.val = tol.num.val, int.as.num = int.as.num, tol.char = tol.char,
+       tol.factor = tol.factor, factor.as.char = factor.as.char, tol.date = tol.date, tol.date.val = tol.date.val,
+       tol.other = tol.other, tol.vars = tol.vars)
 }